# src/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Application": {
				"properties": {
					"id": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"packageName",
					"location"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Device": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"detectedBy": {
						"type": "string"
					},
					"hash": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"detectedBy",
					"hash"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FridaApplication": {
				"properties": {
					"id": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"pid": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"packageName",
					"location",
					"pid"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-frida-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/devices/{deviceId}/adb/getApplications": {
			"get": {
				"operationId": "GetApplications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Application"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices/{deviceId}/adb/install": {
			"post": {
				"operationId": "Install",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"packageName": {
										"type": "string"
									}
								},
								"required": [
									"packageName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/devices/{deviceId}/adb/uninstall": {
			"post": {
				"operationId": "Uninstall",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"packageName": {
										"type": "string"
									}
								},
								"required": [
									"packageName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/devices/{deviceId}/apps": {
			"get": {
				"operationId": "GetApplicationsByDeviceId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Application"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices/{deviceId}/apps/{appName}": {
			"get": {
				"operationId": "GetApplicationByDeviceIdAndAppName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Application"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "appName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices": {
			"get": {
				"operationId": "GetDevices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Device"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/devices/{id}": {
			"get": {
				"operationId": "GetDeviceById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Device"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices/{deviceId}/frida/getApplications": {
			"get": {
				"operationId": "GetApplications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FridaApplication"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices/{deviceId}/frida/inject/{moduleName}": {
			"post": {
				"operationId": "Inject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "deviceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "moduleName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/devices/{deviceId}/frida/spawn": {
			"post": {
				"operationId": "SpawnApplicationByPackageName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"additionalProperties": false,
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"packageName": {
										"type": "string"
									},
									"deviceId": {
										"type": "string"
									}
								},
								"required": [
									"packageName",
									"deviceId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/devices/{deviceId}/frida/spawnAndInject": {
			"post": {
				"operationId": "SpawnAndInject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"b64DecodedFridaScript": {
											"type": "string"
										},
										"packageName": {
											"type": "string"
										},
										"deviceId": {
											"type": "string"
										}
									},
									"required": [
										"b64DecodedFridaScript",
										"packageName",
										"deviceId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"b64EncodedScript": {
										"type": "string"
									},
									"packageName": {
										"type": "string"
									},
									"deviceId": {
										"type": "string"
									}
								},
								"required": [
									"b64EncodedScript",
									"packageName",
									"deviceId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}